+++ packages/openapi/src/payload-config/access-routes/access-schema.ts
