+++ packages/openapi/src/schemas/entity-to-json-schema.ts
