+++ packages/openapi/src/schemas/index.ts
@@ -3,3 +3,4 @@ export * from './basic-components';
 export { default as me } from './me';
 export { createUpsertConfirmationSchema } from './upsert-confirmation';
 export { createPaginatedDocumentSchema } from './paginated-documents';
+export { entityToJSONSchema } from './entity-to-json-schema';
