+++ packages/openapi/src/utils/index.ts
@@ -1,3 +1,2 @@
-export { entityToJSONSchema } from './entity-to-json-schema';
 export { getDescription } from './get-description';
 export { merge } from './merge';
